refactor: extract section DOM building from FileTree into file-tree-section-dom.js
Moves section header construction, chevron toggle, action buttons, and
context menu attachment into a dedicated module. FileTree delegates via
a callbacks object, keeping the component focused on state management.

file-tree.js: 290 → 246 lines
